refactor(http): move McpHttpConfig aggregate to types
